+++ libcaf_core/caf/typed_response_promise.hpp
@@ -5,7 +5,7 @@
 #pragma once
 
 #include "caf/detail/type_list.hpp"
-#include "caf/make_message.hpp"
+#include "caf/message.hpp"
 #include "caf/response_promise.hpp"
 
 #include <type_traits>