+++ libcaf_net/tests/legacy/net/datagram_transport.cpp
@@ -17,7 +17,6 @@
 #include "caf/net/udp_datagram_socket.hpp"
 
 #include "caf/binary_serializer.hpp"
-#include "caf/byte.hpp"
 #include "caf/byte_buffer.hpp"
 #include "caf/detail/socket_sys_includes.hpp"
 #include "caf/make_actor.hpp"