+++ libcaf_net/tests/legacy/net/pipe_socket.cpp
@@ -8,7 +8,6 @@
 
 #include "caf/test/dsl.hpp"
 
-#include "caf/byte.hpp"
 #include "caf/byte_buffer.hpp"
 
 using namespace caf;