+++ CHANGELOG.md
@@ -5,6 +5,14 @@ is based on [Keep a Changelog](https://keepachangelog.com).
 
 ## [Unreleased]
 
+### Removed
+
+- The obsolete CAF types `caf::string_view`, `caf::byte`, `caf::optional`,
+  `caf::replies_to`, and `caf::flow::item_publisher`.
+- The obsolete `operator*` for "combining" two actor handles.
+- All `to_stream` and `to_typed_stream` member functions on actors (they are
+  available on `caf::flow::observable` directly).
+
 ## [0.19.5] - 2024-01-08
 
 ### Added